requirements.txt
Django==4.1.7
requests==2.28.2
python-dotenv==1.0.0
django-environ==0.10.0
openai==0.27.2
django-bootstrap-v5==1.0.11